Use custom numpad for Send Module - Extract custom numpad's Adapter to own class - Replace with custom numpad in SendModule
