+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/pin/PinActivity.kt
@@ -12,7 +12,10 @@ import android.support.v7.widget.GridLayoutManager
 import android.support.v7.widget.LinearLayoutManager
 import android.support.v7.widget.PagerSnapHelper
 import android.support.v7.widget.RecyclerView
-import android.view.*
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import android.view.WindowManager
 import android.view.animation.AnimationUtils
 import android.widget.ImageView
 import android.widget.TextView
@@ -21,12 +24,13 @@ import io.horizontalsystems.bankwallet.R
 import io.horizontalsystems.bankwallet.core.App
 import io.horizontalsystems.bankwallet.core.security.FingerprintAuthenticationDialogFragment
 import io.horizontalsystems.bankwallet.modules.main.MainModule
+import io.horizontalsystems.bankwallet.ui.extensions.NumPadItem
+import io.horizontalsystems.bankwallet.ui.extensions.NumPadItemsAdapter
 import io.horizontalsystems.bankwallet.ui.extensions.SmoothLinearLayoutManager
 import io.horizontalsystems.bankwallet.viewHelpers.DateHelper
 import io.horizontalsystems.bankwallet.viewHelpers.HudHelper
 import kotlinx.android.synthetic.main.activity_pin.*
 import kotlinx.android.synthetic.main.custom_tall_toolbar.*
-import android.view.MotionEvent
 
 
 
@@ -65,23 +69,10 @@ class PinActivity : BaseActivity(), NumPadItemsAdapter.Listener, FingerprintAuth
         viewModel = ViewModelProviders.of(this).get(PinViewModel::class.java)
         viewModel.init(interactionType)
 
-        val numpadAdapter = NumPadItemsAdapter(listOf(
-                NumPadItem(NumPadItemType.NUMBER, 1, ""),
-                NumPadItem(NumPadItemType.NUMBER, 2, "abc"),
-                NumPadItem(NumPadItemType.NUMBER, 3, "def"),
-                NumPadItem(NumPadItemType.NUMBER, 4, "ghi"),
-                NumPadItem(NumPadItemType.NUMBER, 5, "jkl"),
-                NumPadItem(NumPadItemType.NUMBER, 6, "mno"),
-                NumPadItem(NumPadItemType.NUMBER, 7, "pqrs"),
-                NumPadItem(NumPadItemType.NUMBER, 8, "tuv"),
-                NumPadItem(NumPadItemType.NUMBER, 9, "wxyz"),
-                NumPadItem(NumPadItemType.FINGER, 0, "FINGER"),
-                NumPadItem(NumPadItemType.NUMBER, 0, ""),
-                NumPadItem(NumPadItemType.DELETE, 0, "DEL")
-        ), this)
+        val numpadAdapter = NumPadItemsAdapter(this)
 
-        numPadItems.adapter = numpadAdapter
-        numPadItems.layoutManager = GridLayoutManager(this, 3)
+        numPadItemsRecyclerView.adapter = numpadAdapter
+        numPadItemsRecyclerView.layoutManager = GridLayoutManager(this, 3)
 
         viewModel.titleLiveDate.observe(this, Observer { title ->
             title?.let { toolbarTitle.text = getString(it) }
@@ -307,86 +298,5 @@ class PinPageViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
         imgPinMask4.setImageResource(if (length > 3) filledCircle else emptyCircle)
         imgPinMask5.setImageResource(if (length > 4) filledCircle else emptyCircle)
         imgPinMask6.setImageResource(if (length > 5) filledCircle else emptyCircle)
-    }
-}
-
-//NumPad part
-data class NumPadItem(val type: NumPadItemType, val number: Int, val letters: String)
-
-class NumPadItemsAdapter(private val numPadItems: List<NumPadItem>, private val listener: Listener)
-    : RecyclerView.Adapter<RecyclerView.ViewHolder>() {
-
-    interface Listener {
-        fun onItemClick(item: NumPadItem)
-    }
-
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): RecyclerView.ViewHolder {
-        return NumPadItemViewHolder(LayoutInflater.from(parent.context).inflate(R.layout.view_numpad_button, parent, false))
-    }
-
-    override fun getItemCount() = numPadItems.count()
-
-    var showFingerPrintButton = false
-        set(value) {
-            field = value
-            notifyDataSetChanged()
-        }
-
-    override fun onBindViewHolder(holder: RecyclerView.ViewHolder, position: Int) {
-        if (holder is NumPadItemViewHolder) {
-            holder.bind(numPadItems[position], showFingerPrintButton) { listener.onItemClick(numPadItems[position]) }
-        }
-    }
-}
-
-class NumPadItemViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
-
-    private var txtNumber: TextView = itemView.findViewById(R.id.txtNumPadNumber)
-    private var txtLetters: TextView = itemView.findViewById(R.id.txtNumPadText)
-    private var imgBackSpace: ImageView = itemView.findViewById(R.id.imgBackSpace)
-    private var imgFingerprint: ImageView = itemView.findViewById(R.id.imgFingerprint)
-
-
-    fun bind(item: NumPadItem, isFingerprintEnabled: Boolean, onClick: () -> (Unit)) {
-
-        itemView.setOnTouchListener { v, event ->
-            when {
-                event.action == MotionEvent.ACTION_DOWN -> {
-                    onClick.invoke()
-                    v.isPressed = true
-                    true
-                }
-                event.action == MotionEvent.ACTION_UP -> {
-                    v.isPressed = false
-                    true
-                }
-                else -> false
-            }
-        }
-
-        txtNumber.visibility = View.GONE
-        txtLetters.visibility = View.GONE
-        imgBackSpace.visibility = View.GONE
-        imgFingerprint.visibility = View.GONE
-
-        when (item.type) {
-            NumPadItemType.DELETE -> {
-                itemView.background = null
-                imgBackSpace.visibility = View.VISIBLE
-            }
-
-            NumPadItemType.NUMBER -> {
-                txtNumber.visibility = View.VISIBLE
-                txtLetters.visibility = if (item.number == 0) View.GONE else View.VISIBLE
-                txtNumber.text = item.number.toString()
-                txtLetters.text = item.letters
-                itemView.setBackgroundResource(R.drawable.numpad_button_background)
-            }
-
-            NumPadItemType.FINGER -> {
-                itemView.background = null
-                imgFingerprint.visibility = if (isFingerprintEnabled) View.VISIBLE else View.GONE
-            }
-        }
     }
 }
