+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/send/SendBottomSheetFragment.kt
@@ -1,22 +1,31 @@
 package io.horizontalsystems.bankwallet.modules.send
 
+import android.annotation.SuppressLint
 import android.app.Dialog
 import android.arch.lifecycle.Observer
 import android.arch.lifecycle.ViewModelProviders
+import android.content.DialogInterface
 import android.content.Intent
 import android.os.Bundle
+import android.support.design.widget.BottomSheetBehavior
 import android.support.design.widget.BottomSheetDialog
 import android.support.design.widget.BottomSheetDialogFragment
 import android.support.v4.app.FragmentActivity
 import android.support.v4.content.ContextCompat
+import android.support.v7.widget.GridLayoutManager
+import android.support.v7.widget.RecyclerView
 import android.text.Editable
 import android.text.TextUtils
 import android.text.TextWatcher
 import android.util.Log
+import android.view.MotionEvent
 import android.view.View
 import android.widget.*
 import com.google.zxing.integration.android.IntentIntegrator
 import io.horizontalsystems.bankwallet.R
+import io.horizontalsystems.bankwallet.modules.pin.NumPadItemType
+import io.horizontalsystems.bankwallet.ui.extensions.NumPadItem
+import io.horizontalsystems.bankwallet.ui.extensions.NumPadItemsAdapter
 import io.horizontalsystems.bankwallet.viewHelpers.HudHelper
 import io.horizontalsystems.bankwallet.viewHelpers.LayoutHelper
 import io.horizontalsystems.bankwallet.viewHelpers.ValueFormatter
@@ -25,9 +34,10 @@ import io.reactivex.subjects.PublishSubject
 import java.math.BigDecimal
 import java.util.concurrent.TimeUnit
 
-class SendBottomSheetFragment: BottomSheetDialogFragment() {
+class SendBottomSheetFragment : BottomSheetDialogFragment(), NumPadItemsAdapter.Listener {
 
     private lateinit var viewModel: SendViewModel
+    private var amountEditTxt: EditText? = null
 
     private val amountChangeSubject: PublishSubject<Double> = PublishSubject.create()
 
@@ -36,7 +46,7 @@ class SendBottomSheetFragment: BottomSheetDialogFragment() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
-        coin?.let {coin ->
+        coin?.let { coin ->
             activity?.let {
                 viewModel = ViewModelProviders.of(it).get(SendViewModel::class.java)
                 viewModel.init(coin)
@@ -51,11 +61,25 @@ class SendBottomSheetFragment: BottomSheetDialogFragment() {
                 }
     }
 
+    @SuppressLint("ClickableViewAccessibility")
     override fun onCreateDialog(savedInstanceState: Bundle?): Dialog {
 
         val mDialog = activity?.let { BottomSheetDialog(it, R.style.BottomDialog) }
         mDialog?.setContentView(R.layout.fragment_bottom_sheet_pay)
 
+
+        mDialog?.setOnShowListener(object : DialogInterface.OnShowListener {
+            override fun onShow(dialog: DialogInterface?) {
+                val bottomSheet = mDialog.findViewById<FrameLayout>(android.support.design.R.id.design_bottom_sheet)
+                BottomSheetBehavior.from(bottomSheet).state = BottomSheetBehavior.STATE_EXPANDED
+                BottomSheetBehavior.from(bottomSheet).isFitToContents = true
+            }
+        })
+
+
+        val numpadAdapter = NumPadItemsAdapter(this)
+
+        val numpadRecyclerView = mDialog?.findViewById<RecyclerView>(R.id.numPadItemsRecyclerView)
         val hintInfoTxt: TextView? = mDialog?.findViewById(R.id.txtHintInfo)
         val addressTxt: TextView? = mDialog?.findViewById(R.id.txtAddress)
         val addressErrorTxt: TextView? = mDialog?.findViewById(R.id.txtAddressError)
@@ -66,9 +90,10 @@ class SendBottomSheetFragment: BottomSheetDialogFragment() {
         val deleteAddressButton: ImageButton? = mDialog?.findViewById(R.id.btnDeleteAddress)
         val feePrimaryTxt: TextView? = mDialog?.findViewById(R.id.txtFeePrimary)
         val feeSecondaryTxt: TextView? = mDialog?.findViewById(R.id.txtFeeSecondary)
-        val amountEditTxt: EditText? = mDialog?.findViewById(R.id.editTxtAmount)
+        amountEditTxt = mDialog?.findViewById(R.id.editTxtAmount)
         val sendButton: Button? = mDialog?.findViewById(R.id.btnSend)
 
+        amountEditTxt?.showSoftInputOnFocus = false
         sendButton?.isEnabled = false
 
         switchButton?.setOnClickListener { viewModel.delegate.onSwitchClicked() }
@@ -78,6 +103,21 @@ class SendBottomSheetFragment: BottomSheetDialogFragment() {
         sendButton?.setOnClickListener { viewModel.delegate.onSendClicked() }
 
         amountEditTxt?.addTextChangedListener(textChangeListener)
+
+        numpadRecyclerView?.adapter = numpadAdapter
+        numpadRecyclerView?.layoutManager = GridLayoutManager(context, 3)
+
+        //disable BottomSheet dragging in numpad area
+        numpadRecyclerView?.setOnTouchListener { _, event ->
+            when (event?.action) {
+                MotionEvent.ACTION_DOWN, MotionEvent.ACTION_UP -> {
+                    false
+                }
+                else -> {
+                    true
+                }
+            }
+        }
 
         viewModel.switchButtonEnabledLiveData.observe(this, Observer { enabled ->
             enabled?.let { switchButton?.isEnabled = it }
@@ -139,7 +179,7 @@ class SendBottomSheetFragment: BottomSheetDialogFragment() {
 
             if (amountNumber > 0) {
                 amountEditTxt?.setText(BigDecimal.valueOf(amountNumber).toPlainString())
-                amountEditTxt?.setSelection(amountEditTxt.text.length)
+//                amountEditTxt?.setSelection(amountEditTxt.text.length)
             }
         })
 
@@ -207,6 +247,18 @@ class SendBottomSheetFragment: BottomSheetDialogFragment() {
         })
 
         return mDialog as Dialog
+    }
+
+    override fun onItemClick(item: NumPadItem) {
+        when (item.type) {
+            NumPadItemType.NUMBER -> {
+//                amountEditTxt?.append(item.number.toString())
+//                viewModel.delegate.onEnter(item.number.toString())
+            }
+            NumPadItemType.DELETE -> {
+//                viewModel.delegate.onDelete()
+            }
+        }
     }
 
     override fun onActivityResult(requestCode: Int, resultCode: Int, intent: Intent?) {
