+++ app/src/main/java/io/horizontalsystems/bankwallet/ui/extensions/NumPadItemsAdapter.kt
@@ -1,0 +1,107 @@
+package io.horizontalsystems.bankwallet.ui.extensions
+
+import android.support.v7.widget.RecyclerView
+import android.view.LayoutInflater
+import android.view.MotionEvent
+import android.view.View
+import android.view.ViewGroup
+import android.widget.ImageView
+import android.widget.TextView
+import io.horizontalsystems.bankwallet.R
+import io.horizontalsystems.bankwallet.modules.pin.NumPadItemType
+
+class NumPadItemsAdapter(private val listener: Listener) : RecyclerView.Adapter<RecyclerView.ViewHolder>() {
+
+    interface Listener {
+        fun onItemClick(item: NumPadItem)
+    }
+
+    val numPadItems = listOf(
+            NumPadItem(NumPadItemType.NUMBER, 1, ""),
+            NumPadItem(NumPadItemType.NUMBER, 2, "abc"),
+            NumPadItem(NumPadItemType.NUMBER, 3, "def"),
+            NumPadItem(NumPadItemType.NUMBER, 4, "ghi"),
+            NumPadItem(NumPadItemType.NUMBER, 5, "jkl"),
+            NumPadItem(NumPadItemType.NUMBER, 6, "mno"),
+            NumPadItem(NumPadItemType.NUMBER, 7, "pqrs"),
+            NumPadItem(NumPadItemType.NUMBER, 8, "tuv"),
+            NumPadItem(NumPadItemType.NUMBER, 9, "wxyz"),
+            NumPadItem(NumPadItemType.FINGER, 0, "FINGER"),
+            NumPadItem(NumPadItemType.NUMBER, 0, ""),
+            NumPadItem(NumPadItemType.DELETE, 0, "DEL")
+    )
+
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): RecyclerView.ViewHolder {
+        return NumPadItemViewHolder(LayoutInflater.from(parent.context).inflate(R.layout.view_numpad_button, parent, false))
+    }
+
+    override fun getItemCount() = numPadItems.count()
+
+    var showFingerPrintButton = false
+        set(value) {
+            field = value
+            notifyDataSetChanged()
+        }
+
+    override fun onBindViewHolder(holder: RecyclerView.ViewHolder, position: Int) {
+        if (holder is NumPadItemViewHolder) {
+            holder.bind(numPadItems[position], showFingerPrintButton) { listener.onItemClick(numPadItems[position]) }
+        }
+    }
+}
+
+
+data class NumPadItem(val type: NumPadItemType, val number: Int, val letters: String)
+
+
+class NumPadItemViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
+
+    private var txtNumber: TextView = itemView.findViewById(R.id.txtNumPadNumber)
+    private var txtLetters: TextView = itemView.findViewById(R.id.txtNumPadText)
+    private var imgBackSpace: ImageView = itemView.findViewById(R.id.imgBackSpace)
+    private var imgFingerprint: ImageView = itemView.findViewById(R.id.imgFingerprint)
+
+
+    fun bind(item: NumPadItem, isFingerprintEnabled: Boolean, onClick: () -> (Unit)) {
+
+        itemView.setOnTouchListener { v, event ->
+            when {
+                event.action == MotionEvent.ACTION_DOWN -> {
+                    onClick.invoke()
+                    v.isPressed = true
+                    true
+                }
+                event.action == MotionEvent.ACTION_UP -> {
+                    v.isPressed = false
+                    true
+                }
+                else -> false
+            }
+        }
+
+        txtNumber.visibility = View.GONE
+        txtLetters.visibility = View.GONE
+        imgBackSpace.visibility = View.GONE
+        imgFingerprint.visibility = View.GONE
+
+        when (item.type) {
+            NumPadItemType.DELETE -> {
+                itemView.background = null
+                imgBackSpace.visibility = View.VISIBLE
+            }
+
+            NumPadItemType.NUMBER -> {
+                txtNumber.visibility = View.VISIBLE
+                txtLetters.visibility = if (item.number == 0) View.GONE else View.VISIBLE
+                txtNumber.text = item.number.toString()
+                txtLetters.text = item.letters
+                itemView.setBackgroundResource(R.drawable.numpad_button_background)
+            }
+
+            NumPadItemType.FINGER -> {
+                itemView.background = null
+                imgFingerprint.visibility = if (isFingerprintEnabled) View.VISIBLE else View.GONE
+            }
+        }
+    }
+}
