+++ app/src/main/res/layout/activity_pin.xml
@@ -33,14 +33,14 @@
             android:id="@+id/pinPagesRecyclerView"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            app:layout_constraintBottom_toTopOf="@+id/numPadItems"
+            app:layout_constraintBottom_toTopOf="@+id/numPadItemsRecyclerView"
             app:layout_constraintLeft_toLeftOf="parent"
             app:layout_constraintRight_toRightOf="parent"
             app:layout_constraintTop_toTopOf="parent"
             />
 
         <android.support.v7.widget.RecyclerView
-            android:id="@+id/numPadItems"
+            android:id="@+id/numPadItemsRecyclerView"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_marginBottom="32dp"
