+++ app/src/main/res/layout/activity_unlock.xml
@@ -18,7 +18,7 @@
         android:text="@string/Unlock_Page_EnterYourPin"
         android:textColor="@color/grey"
         android:textSize="17sp"
-        app:layout_constraintBottom_toTopOf="@+id/numPadItems"
+        app:layout_constraintBottom_toTopOf="@+id/numPadItemsRecyclerView"
         app:layout_constraintLeft_toLeftOf="parent"
         app:layout_constraintRight_toRightOf="parent"
         app:layout_constraintTop_toTopOf="parent"
@@ -99,7 +99,7 @@
         />
 
     <android.support.v7.widget.RecyclerView
-        android:id="@+id/numPadItems"
+        android:id="@+id/numPadItemsRecyclerView"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginBottom="32dp"
